+++ examples/preemptible_and_regular_instance_templates/simple/main.tf
@@ -24,6 +24,7 @@ provider "google" {
 module "preemptible_and_regular_instance_templates" {
   source          = "../../../modules/preemptible_and_regular_instance_templates"
   subnetwork      = var.subnetwork
+  project_id      = var.project_id
   service_account = var.service_account
   name_prefix     = "pvm-and-regular-simple"
   tags            = var.tags