+++ modules/mig_with_percent/variables.tf
@@ -14,6 +14,12 @@
  * limitations under the License.
  */
 
+variable "project_id" {
+  type        = string
+  description = "The GCP project ID"
+  default     = ""
+}
+
 variable "hostname" {
   description = "Hostname prefix for instances"
   default     = "default"
@@ -67,7 +73,7 @@ variable "update_policy" {
     minimal_action          = string
     type                    = string
   }))
-  default     = []
+  default = []
 }
 
 ##############