+++ CHANGELOG.md
@@ -9,6 +9,10 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Added variable `project_id`. [#25]
+
+### Added
+
 - `preemptible_and_regular_instance_templates` submodule. [#18]
 - `mig_with_percent` submodule. [#18]
 
@@ -42,3 +46,4 @@ project adheres to [Semantic Versioning](http://semver.org/).
 [#14]: https://github.com/terraform-google-modules/terraform-google-vm/pull/14
 [#16]: https://github.com/terraform-google-modules/terraform-google-vm/pull/16
 [#18]: https://github.com/terraform-google-modules/terraform-google-vm/pull/18
+[#25]: https://github.com/terraform-google-modules/terraform-google-vm/pull/25